week1

+++ function-3-3.cpp
@@ -4,9 +4,9 @@ double weightedaverage(int array[], int n) {
     if (n<=0 && n%2 == 0) {
         return 0.0;
     } 
-    double weightAve = 0;
+    double weightAve = 0.0;
     for(int i = 0; i < n; i++) {
-        int sum = 0;
+        double sum = 0;
         for(int j = 0; j < n; j++) {
             if (array[i] == array[j]){
                 sum++;
